IOS-2418 Work with updateView()
